refactor: improve component imports
