+++ src/components/GameResultNotice/NumberIconBall.tsx
